requirements.txt
Django=4.1.13
djongo=4.1.13
pymongo=3.12.3
dnspython>=2.0,<3.0
aiofiles>=0.8,<1.0
asgiref>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
corsheaders>=3.0,<4.0
requests
djangorestframework-simplejwt
djangorestframework